+++ apiclient/googleApiClient.js
@@ -6,14 +6,14 @@ const process = require('process');
 const {authenticate} = require('@google-cloud/local-auth');
 const {google} = require('googleapis');
 const helpers = require("../helpers/helpers");
-const config = require("../config/default");
+const configProvider = require("../config/configurationProvider");
 
 // If modifying these scopes, delete token.json.
 const SCOPES = ['https://www.googleapis.com/auth/calendar.readonly'];
 // The file token.json stores the user's access and refresh tokens, and is
 // created automatically when the authorization flow completes for the first
 // time.
-const TOKEN_PATH = `${config.filePaths.baseFolder}\\gcalToken.json`;
+const TOKEN_PATH = `${configProvider.getConfig().filePaths.baseFolder}\\gcalToken.json`;
 const CREDENTIALS_PATH = path.join(process.cwd(), 'config/credentials.json');
 
 /**