+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Add new command to start time report from google calendar event (description and length)
 - Use past events summaries from google calendar as descriptions on existing commands
+- Implement configuration provider to handle local configurations
 
 ### Changed
 - Bump axios from 1.3.4 to 1.7.5